requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
